pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Saegl <saegl@protonmail.com>", "Sanzhar <sanzhar.baratov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
httpx = "^0.23.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.10.5"}
aerich = "^0.7.1"
pre-commit = "^3.1.1"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.0.0%2Bcpu-cp310-cp310-linux_x86_64.whl"}
numpy = "^1.24.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ruff = "^0.0.241"
pytest = "^7.2.1"
mypy = "^1.0.0"

[tool.aerich]
tortoise_orm = "backend.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
